Store Telegram admin IDs as BIGINT

+++ backend/app/infrastructure_models.py
@@ -4,9 +4,9 @@ import enum
 import uuid
 from datetime import datetime
 
-from sqlalchemy import Boolean, DateTime, Enum, ForeignKey, Integer, String, Text, UniqueConstraint, func
+from sqlalchemy import BigInteger, Boolean, DateTime, Enum, ForeignKey, Integer, String, Text, UniqueConstraint, func
 from sqlalchemy.dialects.postgresql import UUID
-from sqlalchemy.orm import Mapped, mapped_column, relationship
+from sqlalchemy.orm import Mapped, mapped_column
 
 from .db import Base
 
@@ -52,7 +52,7 @@ class ServerOnboardingJob(Base):
     __tablename__ = "server_onboarding_jobs"
 
     id: Mapped[uuid.UUID] = mapped_column(UUID(as_uuid=True), primary_key=True, default=uuid.uuid4)
-    admin_id: Mapped[int] = mapped_column(Integer, index=True)
+    admin_id: Mapped[int] = mapped_column(BigInteger, index=True)
     name: Mapped[str] = mapped_column(String(128))
     country: Mapped[str] = mapped_column(String(128))
     city: Mapped[str] = mapped_column(String(128))
